refactor: Extract parseObservation
